requirements.txt
# Core dependencies
flask>=3.0.0
functions-framework>=3.0.0

# LINE Bot SDK
line-bot-sdk>=3.0.0

# Google APIs
google-generativeai>=0.8.0
gspread>=6.0.0
google-auth>=2.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
